Attempting to clean-up this project.
