+++ _sandbox/colortest.go_1
